HSEARCH-3881 Move filter() methods to more appropriate classes in SearchAggregationBuilder inheritance tree
