+++ engine/src/main/java/org/hibernate/search/engine/search/aggregation/spi/RangeAggregationBuilder.java
@@ -8,9 +8,12 @@ package org.hibernate.search.engine.search.aggregation.spi;
 
 import java.util.Map;
 
+import org.hibernate.search.engine.search.predicate.SearchPredicate;
 import org.hibernate.search.util.common.data.Range;
 
 public interface RangeAggregationBuilder<K> extends SearchAggregationBuilder<Map<Range<K>, Long>> {
+
+	void filter(SearchPredicate filter);
 
 	void range(Range<? extends K> range);
 }
